ai-personality-project/requirements.txt
# Основные зависимости
flask==2.3.2
flask-sqlalchemy==3.0.5
sqlalchemy==2.0.15
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6

# AI/ML зависимости
transformers==4.30.0
torch==2.0.1
numpy==1.24.3

# Утилиты
requests==2.28.2
psutil==5.9.5
python-multipart==0.0.6

# Тестирование
pytest==7.3.1